lambda_agent_final_code/requirements.txt
strands-agents
litellm
python-dotenv